Cygwin: FIFO: designate one reader as owner
Among all the open readers of a FIFO, one is declared to be the owner.
This is the only reader that listens for client connections, and it is
the only one that can read.

Currently the owner is the reader that opened first, and there is no
supported way to give ownership to a different reader.